+++ src/de/danoeh/antennapod/activity/MiroGuideChannelViewActivity.java
@@ -26,8 +26,8 @@ import de.danoeh.antennapod.adapter.MiroGuideItemlistAdapter;
 import de.danoeh.antennapod.dialog.DownloadRequestErrorDialogCreator;
 import de.danoeh.antennapod.feed.Feed;
 import de.danoeh.antennapod.feed.FeedManager;
-import de.danoeh.antennapod.miroguide.con.MiroGuideException;
-import de.danoeh.antennapod.miroguide.con.MiroGuideService;
+import de.danoeh.antennapod.miroguide.conn.MiroGuideException;
+import de.danoeh.antennapod.miroguide.conn.MiroGuideService;
 import de.danoeh.antennapod.miroguide.model.MiroGuideChannel;
 import de.danoeh.antennapod.storage.DownloadRequestException;
 import de.danoeh.antennapod.storage.DownloadRequester;
